Refactor constants & data structure
